Refactor ML Audio API and Remove Compatibility Layer
- Unified the ML Audio API by consolidating functionality into `ml_audio_api.py`, enhancing error handling and fallback mechanisms.
- Removed the deprecated `ml_audio_compatibility.py` file, streamlining the codebase and improving maintainability.
- Updated logging to provide clearer error messages and success confirmations for API integrations.
- Enhanced the system state management to support dynamic capabilities detection and performance metrics tracking.

These changes significantly improve the structure and efficiency of the ML Audio system, ensuring better performance and user experience.
