--- backend/api/ml_audio_compatibility.py
@@ -1,845 +1,0 @@
-#!/usr/bin/env python3
-"""
-ML Audio Compatibility Routes
-Provides backward compatibility for frontend ML audio endpoints
-Routes requests to unified WebSocket handler with enhanced functionality
-"""
-
-from fastapi import APIRouter, HTTPException, Request, Depends
-from pydantic import BaseModel, Field
-from typing import Dict, Any, Optional, List, Tuple
-import logging
-import asyncio
-import base64
-import numpy as np
-from datetime import datetime, timedelta
-import json
-import psutil
-import platform
-import os
-import hashlib
-from collections import deque
-import statistics
-
-logger = logging.getLogger(__name__)
-
-router = APIRouter(prefix="/audio/ml", tags=["ml_audio_compat"])
-
-# Dynamic system state
-class MLAudioSystemState:
-    """Dynamic ML Audio system state manager"""
-    
-    def __init__(self):
-        self.start_time = datetime.now()
-        self.active_streams = {}  # client_id -> stream_info
-        self.total_processed = 0
-        self.last_activity = None
-        self.model_loaded = True
-        self.processing_history = deque(maxlen=1000)  # Last 1000 processing times
-        self.quality_history = deque(maxlen=100)  # Last 100 quality scores
-        self.issue_frequency = {}  # Track issue occurrences
-        self.client_stats = {}  # Per-client statistics
-        self.audio_buffer_stats = {
-            "total_bytes_processed": 0,
-            "average_chunk_size": 0,
-            "peak_processing_rate": 0
-        }
-        self.system_capabilities = self._detect_capabilities()
-        
-    def _detect_capabilities(self) -> Dict[str, bool]:
-        """Dynamically detect system capabilities"""
-        caps = {}
-        
-        # Check for audio processing libraries
-        try:
-            import librosa
-            caps["librosa_available"] = True
-            caps["advanced_audio_analysis"] = True
-        except ImportError:
-            caps["librosa_available"] = False
-            caps["advanced_audio_analysis"] = False
-            
-        # Check for ML frameworks
-        try:
-            import torch
-            caps["pytorch_available"] = True
-            caps["neural_audio_processing"] = True
-        except ImportError:
-            caps["pytorch_available"] = False
-            caps["neural_audio_processing"] = False
-            
-        # Check system resources
-        cpu_count = psutil.cpu_count()
-        total_ram = psutil.virtual_memory().total / (1024**3)  # GB
-        
-        caps["multi_stream_capable"] = cpu_count >= 4
-        caps["high_performance_mode"] = total_ram >= 8
-        caps["gpu_acceleration"] = self._check_gpu()
-        
-        return caps
-    
-    def _check_gpu(self) -> bool:
-        """Check for GPU availability"""
-        try:
-            # Try CUDA
-            import torch
-            return torch.cuda.is_available()
-        except:
-            pass
-            
-        try:
-            # Try Metal (Apple Silicon)
-            import torch
-            return torch.backends.mps.is_available()
-        except:
-            pass
-            
-        return False
-    
-    def get_uptime(self) -> float:
-        """Get system uptime in hours"""
-        return (datetime.now() - self.start_time).total_seconds() / 3600
-    
-    def get_performance_metrics(self) -> Dict[str, Any]:
-        """Calculate real-time performance metrics"""
-        if not self.processing_history:
-            return {
-                "avg_latency_ms": 0,
-                "p95_latency_ms": 0,
-                "p99_latency_ms": 0,
-                "success_rate": 1.0,
-                "throughput_per_second": 0
-            }
-        
-        latencies = list(self.processing_history)
-        sorted_latencies = sorted(latencies)
-        
-        return {
-            "avg_latency_ms": round(statistics.mean(latencies), 2),
-            "p50_latency_ms": round(statistics.median(latencies), 2),
-            "p95_latency_ms": round(sorted_latencies[int(len(sorted_latencies) * 0.95)], 2) if len(sorted_latencies) > 20 else round(max(latencies), 2),
-            "p99_latency_ms": round(sorted_latencies[int(len(sorted_latencies) * 0.99)], 2) if len(sorted_latencies) > 100 else round(max(latencies), 2),
-            "success_rate": round(len([l for l in latencies if l < 1000]) / len(latencies), 3),  # <1s is success
-            "throughput_per_second": round(len(latencies) / max(1, self.get_uptime() * 3600), 2)
-        }
-    
-    def get_quality_insights(self) -> Dict[str, Any]:
-        """Get insights from quality history"""
-        if not self.quality_history:
-            return {
-                "average_quality": 0.85,
-                "quality_trend": "stable",
-                "common_issues": []
-            }
-        
-        qualities = list(self.quality_history)
-        recent_qualities = qualities[-10:] if len(qualities) > 10 else qualities
-        
-        # Determine trend
-        if len(qualities) > 5:
-            first_half_avg = statistics.mean(qualities[:len(qualities)//2])
-            second_half_avg = statistics.mean(qualities[len(qualities)//2:])
-            
-            if second_half_avg > first_half_avg + 0.05:
-                trend = "improving"
-            elif second_half_avg < first_half_avg - 0.05:
-                trend = "degrading"
-            else:
-                trend = "stable"
-        else:
-            trend = "insufficient_data"
-        
-        # Get common issues
-        sorted_issues = sorted(self.issue_frequency.items(), key=lambda x: x[1], reverse=True)
-        common_issues = [issue for issue, _ in sorted_issues[:3]]
-        
-        return {
-            "average_quality": round(statistics.mean(qualities), 3),
-            "recent_average": round(statistics.mean(recent_qualities), 3),
-            "quality_trend": trend,
-            "quality_variance": round(statistics.variance(qualities), 4) if len(qualities) > 1 else 0,
-            "common_issues": common_issues,
-            "best_quality_achieved": round(max(qualities), 3) if qualities else 0,
-            "worst_quality_seen": round(min(qualities), 3) if qualities else 0
-        }
-    
-    def track_processing(self, latency_ms: float):
-        """Track processing metrics"""
-        self.processing_history.append(latency_ms)
-        self.total_processed += 1
-        self.last_activity = datetime.now()
-    
-    def track_quality(self, score: float, issues: List[str]):
-        """Track quality metrics"""
-        self.quality_history.append(score)
-        for issue in issues:
-            self.issue_frequency[issue] = self.issue_frequency.get(issue, 0) + 1
-    
-    def get_client_recommendations(self, client_id: str, user_agent: str = "") -> Dict[str, Any]:
-        """Get personalized recommendations for a client"""
-        # Initialize client stats if new
-        if client_id not in self.client_stats:
-            self.client_stats[client_id] = {
-                "first_seen": datetime.now(),
-                "total_requests": 0,
-                "average_quality": 0.85,
-                "common_issues": []
-            }
-        
-        stats = self.client_stats[client_id]
-        stats["total_requests"] += 1
-        
-        # Dynamic chunk size based on system load
-        active_count = len(self.active_streams)
-        cpu_percent = psutil.cpu_percent(interval=0.1)
-        
-        if cpu_percent > 80 or active_count > 5:
-            recommended_chunk = 1024  # Larger chunks for high load
-        elif cpu_percent < 30 and active_count < 2:
-            recommended_chunk = 256   # Smaller chunks for low latency
-        else:
-            recommended_chunk = 512   # Default
-        
-        # Format recommendation based on client
-        if "chrome" in user_agent.lower():
-            recommended_format = "webm"  # Chrome handles WebM well
-        elif "safari" in user_agent.lower():
-            recommended_format = "wav"   # Safari prefers WAV
-        else:
-            recommended_format = "base64"
-        
-        return {
-            "chunk_size": recommended_chunk,
-            "sample_rate": 16000 if not self.system_capabilities.get("high_performance_mode") else 48000,
-            "format": recommended_format,
-            "enable_preprocessing": cpu_percent < 50,
-            "use_compression": active_count > 3,
-            "adaptive_bitrate": True,
-            "client_profile": {
-                "requests_today": stats["total_requests"],
-                "member_since": stats["first_seen"].isoformat(),
-                "performance_tier": "premium" if stats["total_requests"] > 100 else "standard"
-            }
-        }
-
-# Global system state instance
-system_state = MLAudioSystemState()
-
-# Dynamic configuration generator
-def get_dynamic_config(request: Request) -> Dict[str, Any]:
-    """Generate dynamic configuration based on current system state"""
-    # Get system info
-    cpu_percent = psutil.cpu_percent(interval=0.1)
-    memory = psutil.virtual_memory()
-    
-    # Get network info
-    try:
-        import netifaces
-        interfaces = netifaces.interfaces()
-        network_available = len(interfaces) > 1
-    except:
-        network_available = True
-    
-    # Build dynamic config
-    config = {
-        "model": "silero_vad" if not system_state.system_capabilities.get("neural_audio_processing") else "advanced_neural_vad",
-        "sample_rate": 16000,
-        "chunk_size": 512,
-        "vad_threshold": 0.5,
-        "min_silence_duration_ms": 300,
-        "speech_pad_ms": 30,
-        "features": {
-            "voice_activity_detection": True,
-            "noise_suppression": True,
-            "echo_cancellation": cpu_percent < 70,  # Disable if high CPU
-            "automatic_gain_control": True,
-            "speech_enhancement": memory.percent < 80,  # Disable if low memory
-            "background_noise_reduction": True,
-            "advanced_processing": system_state.system_capabilities.get("advanced_audio_analysis", False)
-        },
-        "supported_formats": ["base64", "raw", "wav", "webm"],
-        "max_audio_length_seconds": 60 if memory.percent < 80 else 30,
-        "backend_available": True,
-        "websocket_endpoint": os.getenv("WEBSOCKET_ENDPOINT", "/ws"),
-        "legacy_endpoints_deprecated": True,
-        "performance": system_state.get_performance_metrics(),
-        "advanced_features": {
-            "emotion_detection": system_state.system_capabilities.get("neural_audio_processing", False),
-            "speaker_diarization": system_state.system_capabilities.get("neural_audio_processing", False) and cpu_percent < 50,
-            "language_detection": True,
-            "transcription_available": system_state.system_capabilities.get("librosa_available", False),
-            "real_time_enhancement": cpu_percent < 60
-        },
-        "system_status": {
-            "cpu_usage": cpu_percent,
-            "memory_usage": memory.percent,
-            "network_available": network_available,
-            "uptime_hours": round(system_state.get_uptime(), 2),
-            "platform": platform.system(),
-            "capabilities": system_state.system_capabilities
-        }
-    }
-    
-    return config
-
-class AudioPrediction(BaseModel):
-    audio_data: Optional[str] = Field(None, description="Base64 encoded audio data")
-    features: Optional[Dict[str, Any]] = Field(None, description="Optional audio features")
-    format: Optional[str] = Field("base64", description="Audio format")
-    sample_rate: Optional[int] = Field(16000, description="Sample rate in Hz")
-    duration_ms: Optional[int] = Field(None, description="Audio duration in milliseconds")
-    client_id: Optional[str] = Field(None, description="Client identifier")
-
-class AudioAnalysis(BaseModel):
-    prediction: str
-    confidence: float
-    issues: List[str]
-    recommendations: List[str]
-    audio_quality: Dict[str, Any]
-    processing_time_ms: float
-    detailed_analysis: Dict[str, Any]
-
-def analyze_audio_quality(audio_data: str, format: str = "base64", client_id: Optional[str] = None) -> Dict[str, Any]:
-    """Analyze audio quality with dynamic scoring"""
-    # Base quality calculation using data characteristics
-    if audio_data:
-        # Use data length and format to influence score
-        data_length = len(audio_data)
-        
-        # Hash-based pseudo-randomness for consistent results per audio
-        data_hash = int(hashlib.md5(audio_data.encode()).hexdigest()[:8], 16)
-        base_score = 0.7 + (data_hash % 30) / 100  # 0.7-1.0 range
-        
-        # Adjust based on format
-        format_multipliers = {
-            "wav": 1.05,
-            "webm": 1.02,
-            "base64": 1.0,
-            "raw": 0.95
-        }
-        
-        quality_score = min(1.0, base_score * format_multipliers.get(format, 1.0))
-        
-        # Length-based adjustments
-        if data_length < 1000:  # Very short audio
-            quality_score *= 0.9
-        elif data_length > 100000:  # Long audio
-            quality_score *= 0.95
-    else:
-        quality_score = 0.0
-    
-    # Determine quality level
-    quality_level = "unusable"
-    for level_name, threshold in [
-        ("excellent", 0.9),
-        ("good", 0.7),
-        ("fair", 0.5),
-        ("poor", 0.3)
-    ]:
-        if quality_score >= threshold:
-            quality_level = level_name
-            break
-    
-    # Dynamic issue detection based on score and system state
-    issues = []
-    if quality_score < 0.95:
-        possible_issues = {
-            "background_noise": quality_score < 0.9 and system_state.issue_frequency.get("background_noise", 0) > 5,
-            "echo": quality_score < 0.85 and "echo" in str(audio_data)[:100],  # Simple heuristic
-            "clipping": quality_score < 0.8 and data_length % 1000 < 100,  # Pattern detection
-            "low_volume": quality_score < 0.75,
-            "distortion": quality_score < 0.7,
-            "interference": quality_score < 0.6 and datetime.now().second % 3 == 0,  # Time-based
-            "codec_issues": format == "webm" and quality_score < 0.8
-        }
-        
-        issues = [issue for issue, condition in possible_issues.items() if condition]
-    
-    # Calculate detailed metrics
-    snr_base = 15 + quality_score * 35  # 15-50 dB range
-    snr = round(snr_base + (datetime.now().microsecond % 1000) / 200, 1)  # Add variation
-    
-    result = {
-        "score": round(quality_score, 3),
-        "level": quality_level,
-        "description": f"{quality_level.capitalize()} audio quality detected",
-        "issues_detected": issues,
-        "signal_to_noise_ratio": snr,
-        "peak_amplitude": round(0.6 + quality_score * 0.35, 2),  # 0.6-0.95 range
-        "rms_level": round(-25 + quality_score * 20, 1),  # -25 to -5 dB range
-        "frequency_response": "20Hz-20kHz" if quality_score > 0.8 else "50Hz-15kHz",
-        "dynamic_range": round(60 + quality_score * 30, 1),  # 60-90 dB
-        "thd_percent": round((1 - quality_score) * 5, 2)  # Total harmonic distortion
-    }
-    
-    # Track quality for insights
-    system_state.track_quality(quality_score, issues)
-    
-    return result
-
-def generate_recommendations(issues: List[str], quality_score: float, client_id: Optional[str] = None) -> List[str]:
-    """Generate intelligent, context-aware recommendations"""
-    recommendations = []
-    
-    # Get quality insights
-    insights = system_state.get_quality_insights()
-    
-    # Issue-specific solutions with context
-    issue_solutions = {
-        "background_noise": [
-            "Enable AI-powered noise suppression in your audio settings",
-            "Use a directional microphone to reduce ambient noise",
-            "Record in a quieter environment or use acoustic treatment"
-        ],
-        "echo": [
-            "Use headphones to prevent audio feedback",
-            "Reduce speaker volume or increase distance from microphone",
-            "Enable echo cancellation in your audio processing pipeline"
-        ],
-        "clipping": [
-            "Reduce input gain by 10-15%",
-            "Enable automatic gain control (AGC)",
-            "Check for audio limiter settings in your recording software"
-        ],
-        "low_volume": [
-            "Increase microphone gain gradually until optimal",
-            "Position microphone 6-12 inches from sound source",
-            "Check system audio boost settings"
-        ],
-        "distortion": [
-            "Lower the input sensitivity",
-            "Check cable connections for damage",
-            "Update audio drivers to latest version"
-        ],
-        "interference": [
-            "Move away from electromagnetic sources",
-            "Use shielded audio cables",
-            "Check for ground loop issues"
-        ],
-        "codec_issues": [
-            "Try using WAV format for better quality",
-            "Update your browser or audio codec",
-            "Check bitrate settings in your encoder"
-        ]
-    }
-    
-    # Add relevant solutions
-    for issue in issues:
-        if issue in issue_solutions:
-            # Pick most relevant solution based on frequency
-            solutions = issue_solutions[issue]
-            if system_state.issue_frequency.get(issue, 0) > 10:
-                recommendations.append(solutions[0])  # Most likely solution
-            else:
-                recommendations.append(solutions[len(recommendations) % len(solutions)])
-    
-    # Quality-based recommendations
-    if quality_score < 0.6:
-        recommendations.append("Consider upgrading to a professional audio interface")
-    elif quality_score < 0.7 and insights["quality_trend"] == "degrading":
-        recommendations.append("Audio quality is declining - check microphone condition")
-    elif quality_score < 0.8 and not issues:
-        recommendations.append("Enable audio enhancement features for optimal quality")
-    
-    # Trend-based recommendations
-    if insights["quality_trend"] == "improving":
-        recommendations.append("Your audio quality is improving! Keep current settings")
-    elif insights["quality_variance"] > 0.01:
-        recommendations.append("Audio quality is inconsistent - check for environmental changes")
-    
-    # System-based recommendations
-    if system_state.system_capabilities.get("advanced_audio_analysis") and quality_score < 0.9:
-        recommendations.append("Advanced audio analysis available - enable for better processing")
-    
-    # Limit recommendations
-    return recommendations[:3] if recommendations else ["Audio quality is optimal"]
-
-@router.get("/config")
-async def get_ml_config(request: Request):
-    """Get dynamic ML audio configuration"""
-    logger.info(f"ML Audio config requested from {request.client.host if request.client else 'unknown'}")
-    
-    # Generate dynamic config
-    config = get_dynamic_config(request)
-    
-    # Add client-specific info
-    client_id = f"{request.client.host}_{request.headers.get('user-agent', 'unknown')}" if request.client else "default"
-    
-    config["client_info"] = {
-        "ip": request.client.host if request.client else "unknown",
-        "user_agent": request.headers.get("user-agent", "unknown"),
-        "recommended_settings": system_state.get_client_recommendations(
-            client_id, 
-            request.headers.get("user-agent", "")
-        )
-    }
-    
-    # Add quality insights
-    config["quality_insights"] = system_state.get_quality_insights()
-    
-    return config
-
-@router.post("/predict")
-async def predict_audio_issue(data: AudioPrediction, request: Request):
-    """Enhanced audio prediction with detailed analysis"""
-    start_time = datetime.now()
-    
-    # Get client ID
-    client_id = data.client_id or f"{request.client.host if request.client else 'anonymous'}_{datetime.now().timestamp()}"
-    
-    logger.info(f"ML Audio prediction from {client_id} - format: {data.format}, size: {len(data.audio_data) if data.audio_data else 0}")
-    
-    # Analyze audio quality
-    quality_analysis = analyze_audio_quality(
-        data.audio_data or "",
-        data.format or "base64",
-        client_id
-    )
-    
-    # Dynamic prediction based on analysis
-    confidence_boost = len(system_state.quality_history) / 1000  # Increase confidence with more data
-    
-    if quality_analysis["score"] > 0.9:
-        prediction = "excellent"
-        confidence = min(0.99, 0.95 + confidence_boost)
-    elif quality_analysis["score"] > 0.7:
-        prediction = "normal"
-        confidence = min(0.95, 0.85 + confidence_boost)
-    elif quality_analysis["score"] > 0.5:
-        prediction = "degraded"
-        confidence = min(0.85, 0.75 + confidence_boost)
-    else:
-        prediction = "poor"
-        confidence = min(0.75, 0.65 + confidence_boost)
-    
-    # Generate intelligent recommendations
-    recommendations = generate_recommendations(
-        quality_analysis["issues_detected"],
-        quality_analysis["score"],
-        client_id
-    )
-    
-    # Calculate processing time
-    processing_time = (datetime.now() - start_time).total_seconds() * 1000
-    system_state.track_processing(processing_time)
-    
-    # Build comprehensive response
-    response = AudioAnalysis(
-        prediction=prediction,
-        confidence=round(confidence, 3),
-        issues=quality_analysis["issues_detected"],
-        recommendations=recommendations,
-        audio_quality=quality_analysis,
-        processing_time_ms=round(processing_time, 2),
-        detailed_analysis={
-            "vad_detected": quality_analysis["score"] > 0.3,
-            "speech_segments": [] if quality_analysis["score"] < 0.5 else [
-                {
-                    "start_ms": 0, 
-                    "end_ms": data.duration_ms or 1000,
-                    "confidence": confidence,
-                    "energy_level": quality_analysis["peak_amplitude"]
-                }
-            ],
-            "noise_profile": {
-                "type": "complex" if len(quality_analysis["issues_detected"]) > 2 else 
-                       "ambient" if "background_noise" in quality_analysis["issues_detected"] else "clean",
-                "level_db": quality_analysis["rms_level"],
-                "frequency_mask": [bool(i % 2) for i in range(8)]  # Frequency band mask
-            },
-            "spectral_features": {
-                "centroid_hz": 2000 + quality_analysis["score"] * 2000,
-                "rolloff_hz": 5000 + quality_analysis["score"] * 5000,
-                "flux": round(0.1 + (1 - quality_analysis["score"]) * 0.3, 3),
-                "mfcc_available": system_state.system_capabilities.get("librosa_available", False)
-            },
-            "format_info": {
-                "input_format": data.format,
-                "sample_rate": data.sample_rate,
-                "duration_ms": data.duration_ms,
-                "estimated_bitrate": 128 if data.format == "webm" else 256
-            },
-            "system_load": {
-                "current_streams": len(system_state.active_streams),
-                "processing_capacity": f"{100 - psutil.cpu_percent():.1f}%",
-                "queue_depth": 0  # Would be actual queue depth in production
-            },
-            "enhancement_applied": quality_analysis["score"] < 0.8,
-            "migration_note": "For real-time processing, please use WebSocket at " + os.getenv("WEBSOCKET_ENDPOINT", "/ws")
-        }
-    )
-    
-    return response.dict()
-
-@router.get("/status")
-async def get_ml_audio_status():
-    """Get comprehensive ML audio system status"""
-    # Get real system metrics
-    cpu_percent = psutil.cpu_percent(interval=0.1)
-    memory = psutil.virtual_memory()
-    disk = psutil.disk_usage('/')
-    
-    # Get process info
-    process = psutil.Process()
-    process_info = {
-        "memory_mb": round(process.memory_info().rss / 1024 / 1024, 1),
-        "cpu_percent": process.cpu_percent(interval=0.1),
-        "threads": process.num_threads(),
-        "open_files": len(process.open_files()) if hasattr(process, "open_files") else 0
-    }
-    
-    # Network statistics
-    try:
-        net_io = psutil.net_io_counters()
-        network_stats = {
-            "bytes_sent_mb": round(net_io.bytes_sent / 1024 / 1024, 1),
-            "bytes_recv_mb": round(net_io.bytes_recv / 1024 / 1024, 1),
-            "packets_dropped": net_io.dropin + net_io.dropout
-        }
-    except:
-        network_stats = {"status": "unavailable"}
-    
-    return {
-        "status": "operational" if cpu_percent < 90 and memory.percent < 90 else "degraded",
-        "model_loaded": system_state.model_loaded,
-        "websocket_available": True,
-        "websocket_endpoint": os.getenv("WEBSOCKET_ENDPOINT", "/ws"),
-        "system_health": {
-            "cpu_usage_percent": cpu_percent,
-            "memory_usage_mb": round(memory.used / 1024 / 1024, 1),
-            "memory_percent": memory.percent,
-            "disk_usage_percent": disk.percent,
-            "active_streams": len(system_state.active_streams),
-            "total_processed_today": system_state.total_processed,
-            "last_activity": system_state.last_activity.isoformat() if system_state.last_activity else None,
-            "uptime_hours": round(system_state.get_uptime(), 2),
-            "process_info": process_info,
-            "network": network_stats
-        },
-        "performance_metrics": system_state.get_performance_metrics(),
-        "quality_insights": system_state.get_quality_insights(),
-        "capabilities": {
-            "real_time_processing": True,
-            "batch_processing": True,
-            "multi_language": True,
-            "noise_cancellation": True,
-            "echo_suppression": cpu_percent < 80,
-            "advanced_features": system_state.system_capabilities
-        },
-        "model_info": {
-            "name": "silero_vad" if not system_state.system_capabilities.get("neural_audio_processing") else "advanced_neural_vad",
-            "version": "4.0",
-            "last_updated": "2024-01-15",
-            "accuracy": 0.97,
-            "framework": "PyTorch" if system_state.system_capabilities.get("pytorch_available") else "NumPy"
-        },
-        "audio_buffer_stats": system_state.audio_buffer_stats,
-        "issue_statistics": dict(sorted(system_state.issue_frequency.items(), key=lambda x: x[1], reverse=True)[:5]),
-        "recommendations": {
-            "system": "Upgrade to 16GB RAM for optimal performance" if memory.total < 16 * 1024**3 else "System resources optimal",
-            "configuration": "Enable GPU acceleration" if not system_state.system_capabilities.get("gpu_acceleration") else "GPU acceleration active"
-        },
-        "api_version": "2.0",
-        "legacy_notice": "This endpoint provides compatibility. For best performance, migrate to unified WebSocket at " + os.getenv("WEBSOCKET_ENDPOINT", "/ws")
-    }
-
-@router.post("/analyze")
-async def analyze_audio_detailed(data: AudioPrediction, request: Request):
-    """Detailed audio analysis with spectral features"""
-    logger.info("Detailed audio analysis requested")
-    
-    # Get client ID
-    client_id = data.client_id or f"{request.client.host if request.client else 'anonymous'}"
-    
-    # Perform comprehensive analysis
-    quality = analyze_audio_quality(data.audio_data or "", data.format or "base64", client_id)
-    
-    # Generate spectral features (simulated but realistic)
-    base_centroid = 1500 + quality["score"] * 2500
-    time_factor = datetime.now().microsecond / 1000000  # 0-1 range
-    
-    spectral_features = {
-        "spectral_centroid": round(base_centroid + time_factor * 500, 1),
-        "spectral_rolloff": round(base_centroid * 2.5, 1),
-        "spectral_bandwidth": round(1000 + quality["score"] * 1000, 1),
-        "zero_crossing_rate": round(0.05 + quality["score"] * 0.1 + time_factor * 0.05, 3),
-        "spectral_contrast": [round(20 + i * 5 - quality["score"] * 10, 1) for i in range(7)],
-        "mfcc_coefficients": [round(np.random.randn() * 10, 2) for _ in range(13)] if data.audio_data else None,
-        "chroma_features": [round(abs(np.random.randn() * 0.5), 3) for _ in range(12)] if data.audio_data else None,
-        "tempo_bpm": round(60 + quality["score"] * 60 + time_factor * 20) if quality["score"] > 0.5 else None
-    }
-    
-    # Energy analysis
-    energy_features = {
-        "total_energy": round(quality["peak_amplitude"] ** 2, 3),
-        "energy_entropy": round(0.5 + (1 - quality["score"]) * 0.3, 3),
-        "short_time_energy": [round(0.5 + np.random.random() * 0.4, 2) for _ in range(10)],
-        "energy_variance": round(0.1 + (1 - quality["score"]) * 0.2, 3)
-    }
-    
-    return {
-        "analysis": quality,
-        "spectral_features": spectral_features,
-        "energy_features": energy_features,
-        "temporal_features": {
-            "attack_time_ms": round(10 + (1 - quality["score"]) * 40, 1),
-            "decay_time_ms": round(50 + quality["score"] * 100, 1),
-            "sustain_level": round(quality["peak_amplitude"] * 0.7, 2),
-            "release_time_ms": round(100 + quality["score"] * 200, 1)
-        },
-        "perceptual_features": {
-            "loudness_lufs": round(-23 + quality["score"] * 10, 1),
-            "sharpness": round(0.5 + quality["score"] * 0.3, 2),
-            "roughness": round((1 - quality["score"]) * 0.5, 2),
-            "warmth": round(0.3 + quality["score"] * 0.4, 2)
-        },
-        "recommendations": generate_recommendations(quality["issues_detected"], quality["score"], client_id),
-        "processing_capabilities": system_state.system_capabilities,
-        "advanced_analysis_available": system_state.system_capabilities.get("advanced_audio_analysis", False),
-        "websocket_required": "For real-time feature extraction, use " + os.getenv("WEBSOCKET_ENDPOINT", "/ws")
-    }
-
-@router.get("/models")
-async def list_available_models():
-    """List dynamically available ML audio models"""
-    models = []
-    
-    # Base model (always available)
-    models.append({
-        "id": "silero_vad",
-        "name": "Silero Voice Activity Detection",
-        "version": "4.0",
-        "status": "loaded" if system_state.model_loaded else "available",
-        "capabilities": ["vad", "speech_detection", "noise_filtering"],
-        "resource_requirements": {"cpu": "low", "memory": "128MB", "gpu": "optional"},
-        "performance": {"latency_ms": 10, "accuracy": 0.95}
-    })
-    
-    # Conditionally available models
-    if system_state.system_capabilities.get("neural_audio_processing"):
-        models.append({
-            "id": "whisper_base",
-            "name": "OpenAI Whisper Base",
-            "version": "1.0",
-            "status": "available",
-            "capabilities": ["transcription", "language_detection", "timestamps"],
-            "resource_requirements": {"cpu": "medium", "memory": "1GB", "gpu": "recommended"},
-            "performance": {"latency_ms": 500, "accuracy": 0.98},
-            "languages": ["en", "es", "fr", "de", "ja", "ko", "zh"]
-        })
-        
-        models.append({
-            "id": "neural_enhancer",
-            "name": "Neural Audio Enhancement",
-            "version": "2.0",
-            "status": "available",
-            "capabilities": ["noise_reduction", "echo_cancellation", "quality_enhancement"],
-            "resource_requirements": {"cpu": "high", "memory": "512MB", "gpu": "required"},
-            "performance": {"latency_ms": 50, "improvement": "15-25dB SNR"}
-        })
-    
-    if system_state.system_capabilities.get("librosa_available"):
-        models.append({
-            "id": "spectral_analyzer",
-            "name": "Advanced Spectral Analyzer",
-            "version": "1.5",
-            "status": "loaded",
-            "capabilities": ["spectral_analysis", "feature_extraction", "music_analysis"],
-            "resource_requirements": {"cpu": "low", "memory": "256MB", "gpu": "optional"},
-            "performance": {"latency_ms": 20, "features": 50}
-        })
-    
-    # RNNoise (lightweight, always available)
-    models.append({
-        "id": "rnnoise",
-        "name": "RNNoise Suppression",
-        "version": "0.9",
-        "status": "loaded",
-        "capabilities": ["noise_suppression", "real_time_processing"],
-        "resource_requirements": {"cpu": "minimal", "memory": "64MB", "gpu": "not_required"},
-        "performance": {"latency_ms": 5, "suppression_db": 20}
-    })
-    
-    return {
-        "models": models,
-        "active_model": "silero_vad" if not system_state.system_capabilities.get("neural_audio_processing") else "neural_enhancer",
-        "gpu_available": system_state.system_capabilities.get("gpu_acceleration", False),
-        "total_models": len(models),
-        "loaded_models": len([m for m in models if m["status"] == "loaded"]),
-        "system_capabilities": system_state.system_capabilities,
-        "model_switching_enabled": True,
-        "load_model_endpoint": "/audio/ml/models/load",
-        "benchmark_endpoint": "/audio/ml/models/benchmark"
-    }
-
-@router.post("/stream/start")
-async def start_audio_stream(request: Request):
-    """Start a new audio stream session"""
-    client_id = f"{request.client.host if request.client else 'anonymous'}_{datetime.now().timestamp()}"
-    
-    # Register stream
-    system_state.active_streams[client_id] = {
-        "started_at": datetime.now(),
-        "processed_chunks": 0,
-        "total_bytes": 0,
-        "quality_scores": []
-    }
-    
-    logger.info(f"Started audio stream for client {client_id}")
-    
-    return {
-        "stream_id": client_id,
-        "status": "active",
-        "recommended_config": system_state.get_client_recommendations(client_id, request.headers.get("user-agent", "")),
-        "server_time": datetime.now().isoformat(),
-        "max_duration_seconds": 300,  # 5 minutes max
-        "websocket_upgrade_available": True,
-        "websocket_endpoint": os.getenv("WEBSOCKET_ENDPOINT", "/ws")
-    }
-
-@router.post("/stream/{stream_id}/stop")
-async def stop_audio_stream(stream_id: str):
-    """Stop an active audio stream"""
-    if stream_id in system_state.active_streams:
-        stream_info = system_state.active_streams[stream_id]
-        duration = (datetime.now() - stream_info["started_at"]).total_seconds()
-        
-        # Calculate statistics
-        avg_quality = statistics.mean(stream_info["quality_scores"]) if stream_info["quality_scores"] else 0
-        
-        # Remove from active streams
-        del system_state.active_streams[stream_id]
-        
-        logger.info(f"Stopped audio stream {stream_id}")
-        
-        return {
-            "stream_id": stream_id,
-            "status": "stopped",
-            "duration_seconds": round(duration, 2),
-            "chunks_processed": stream_info["processed_chunks"],
-            "total_bytes": stream_info["total_bytes"],
-            "average_quality": round(avg_quality, 3),
-            "final_report": {
-                "quality_trend": "stable" if len(stream_info["quality_scores"]) < 5 else 
-                               "improving" if stream_info["quality_scores"][-1] > stream_info["quality_scores"][0] else "degrading",
-                "recommendations": generate_recommendations([], avg_quality, stream_id) if avg_quality > 0 else []
-            }
-        }
-    else:
-        raise HTTPException(status_code=404, detail=f"Stream {stream_id} not found")
-
-# Health check endpoint
-@router.get("/health")
-async def ml_audio_health():
-    """Quick health check for ML audio system"""
-    return {
-        "status": "healthy",
-        "timestamp": datetime.now().isoformat(),
-        "active_streams": len(system_state.active_streams),
-        "uptime_hours": round(system_state.get_uptime(), 2),
-        "last_activity": system_state.last_activity.isoformat() if system_state.last_activity else None,
-        "models_available": system_state.model_loaded
-    }
-
-# Note: WebSocket endpoint ws://localhost:8000/audio/ml/stream is handled by 
-# the main.py ML Audio WebSocket compatibility endpoint
